Changes to backend
Addad libraries and changes to ApiRunner
