+++ .idea/libraries/Maven__com_sparkjava_spark_core_2_9_3.xml
@@ -1,0 +1,13 @@
+<component name="libraryTable">
+  <library name="Maven: com.sparkjava:spark-core:2.9.3">
+    <CLASSES>
+      <root url="jar://$MAVEN_REPOSITORY$/com/sparkjava/spark-core/2.9.3/spark-core-2.9.3.jar!/" />
+    </CLASSES>
+    <JAVADOC>
+      <root url="jar://$MAVEN_REPOSITORY$/com/sparkjava/spark-core/2.9.3/spark-core-2.9.3-javadoc.jar!/" />
+    </JAVADOC>
+    <SOURCES>
+      <root url="jar://$MAVEN_REPOSITORY$/com/sparkjava/spark-core/2.9.3/spark-core-2.9.3-sources.jar!/" />
+    </SOURCES>
+  </library>
+</component>
