+++ .idea/modules.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="ProjectModuleManager">
+    <modules>
+      <module fileurl="file://$PROJECT_DIR$/src/Webbtjanster-projekt.iml" filepath="$PROJECT_DIR$/src/Webbtjanster-projekt.iml" />
+      <module fileurl="file://$PROJECT_DIR$/java-backend.iml" filepath="$PROJECT_DIR$/java-backend.iml" />
+    </modules>
+  </component>
+</project>
